// src/bounded-contexts/inventory/domain/factory/inventory_factory/interface.ts
export class InventoryCreationError extends Schema.TaggedError<InventoryCreationError>()('InventoryCreationError', {
  reason: Schema.String,
  invalidFields: Schema.Array(Schema.String),
  context: Schema.Record({ key: Schema.String, value: Schema.Unknown }).pipe(Schema.optional),
}) {}
// ...
export type InventoryType = 'player' | 'creative' | 'survival' | 'spectator' | 'adventure'